refactor: move decorator to separate module
